requirements.txt
Django
pillow
pytesseract
django-widget-tweaks
